Refactored permissions into multiple files.
